clean up all rubocop violations; add exclusions to rubocop.yml
