+++ app/search_builders/concerns/newspaper_works/highlight_search_params.rb
@@ -2,7 +2,6 @@ module NewspaperWorks
   # add highlighting on _stored_ full text field if this is a keyword search
   # can be added to default_processor_chain in a SearchBuilder class
   module HighlightSearchParams
-
     # add highlights on full text field, if there is a keyword query
     def highlight_search_params(solr_parameters = {})
       return unless solr_parameters[:q] || solr_parameters[:all_fields]
