+++ lib/generators/newspaper_works/templates/newspaper_works_helper.rb
