+++ spec/model_shared.rb
@@ -70,7 +70,6 @@ RSpec.shared_examples 'a work and PCDM object' do
   end
 end
 
-# rubocop:disable Metrics/MethodLength
 def model_fixtures(target_type)
   # set up graph of related objects, setting membership on aggregating
   # parents via members setter method, per PCDM Profile for Newspapers.
@@ -133,4 +132,3 @@ def model_fixtures(target_type)
   # return multiple objects as needed for testing
   return [page1, page2] if target_type == :newspaper_pages
 end
-# rubocop:enable Metrics/MethodLength
